Rename style -> styled
